requirements.txt
botcity-framework-core>=0.4.0,<1.0
botcity-framework-web>=0.7.0,<1.0
botcity-maestro-sdk>=0.2.0,<1.0
botcity-excel-plugin==0.1.1
